+++ iree/turbine/kernel/wave/codegen/handlers.py
@@ -62,7 +62,6 @@ from ...ops.wave_ops import (
     exp2,
     extract,
     extract_slice,
-    gather_to_lds,
     ge,
     get_custom,
     get_result,
@@ -1723,83 +1722,3 @@ def handle_reshape(emitter: WaveEmitter, node: fx.Node):
         [1],
     )
     emitter.bind_node_proxy(node, IRProxyValue(slice))
-
-
-@handle_op(gather_to_lds)
-def handle_gather_to_lds(emitter: WaveEmitter, node: fx.Node):
-    try:
-        (
-            src,
-            src_idx,
-            src_type,
-            dst,
-            dst_idx,
-            dst_type,
-            src_mapping,
-            dst_mapping,
-            elements_per_thread,
-        ) = node.args
-    except ValueError as e:
-        raise ValidationError("Malformed arguments") from e
-
-    element_type = IrType.parse(src_type.dtype.ir_type_asm())
-
-    src = cast_py_value(emitter, src)
-    dst = cast_py_value(emitter, dst)
-    src_data_type = get_type_or_element_type(src.ir_value.type)
-    dst_data_type = get_type_or_element_type(dst.ir_value.type)
-
-    if not (
-        MemRefType.isinstance(src.ir_value.type)
-        and MemRefType.isinstance(dst.ir_value.type)
-    ):
-        op = get_custom(node)
-        raise ValidationError(
-            f"Expected src and dst to be of Memref type for\n"
-            f"{op}\nGot\n"
-            f"src: {src.ir_value.type}\n"
-            f"dst: {dst.ir_value.type}\n"
-        )
-
-    if src_data_type != dst_data_type:
-        op = get_custom(node)
-        raise ValidationError(
-            f"Expected src and dst to have same data type for\n"
-            f"{op}\nGot\n"
-            f"src: {src_data_type} vs dst: {dst_data_type}\n"
-        )
-
-    src = src.ir_value
-    dst = dst.ir_value
-
-    src_index_transformed, dst_index_transformed = src_idx, dst_idx
-    if src_mapping:
-        src_index_transformed = transform_index_on_mapping(
-            src_mapping, src_type.symbolic_shape, src_idx
-        )
-    if dst_mapping:
-        dst_index_transformed = transform_index_on_mapping(
-            dst_mapping, dst_type.symbolic_shape, dst_idx
-        )
-    src_keys = list(src_index_transformed.keys())
-    src_fastest_dim = get_fastest_index(src_idx)
-    dst_keys = list(dst_index_transformed.keys())
-    dst_fastest_dim = get_fastest_index(dst_idx)
-    for i in range(elements_per_thread):
-        new_src_index = copy.deepcopy(src_index_transformed)
-        src_key = src_keys[src_fastest_dim]
-        new_src_index[src_key].start += i
-        src_index_transformed_ = _build_start_indices(emitter, new_src_index)
-        new_dst_index = copy.deepcopy(dst_index_transformed)
-        dst_key = dst_keys[dst_fastest_dim]
-        new_dst_index[dst_key].start += i
-        dst_index_transformed_ = _build_start_indices(emitter, new_dst_index)
-        amdgpu_d.gather_to_lds(
-            src=src,
-            src_indices=src_index_transformed_,
-            dst=dst,
-            dst_indices=dst_index_transformed_,
-            transfer_type=element_type,
-        )
-
-    amdgpu_d.lds_barrier()
